+++ examples/train/rft/math.json
