+++ swift/cli/sample.py
@@ -1,5 +1,5 @@
 # Copyright (c) Alibaba, Inc. and its affiliates.
-from swift.experimental.sampling.sampling import sampling_main
+from swift.llm.sampling import sampling_main
 
 if __name__ == '__main__':
     sampling_main()
