+++ swift/llm/sampling/__init__.py
@@ -1,0 +1,1 @@
+from .sampling import sampling_main
