+++ swift/llm/sampling/base.py
@@ -1,7 +1,7 @@
 from swift.plugin.orm import orms
 from swift.plugin.prm import prms
 from swift.utils import get_logger
-from .sampling_args import SamplingArguments
+from swift.llm.argument.sampling_args import SamplingArguments
 
 logger = get_logger()
