+++ swift/llm/sampling/utils.py
