+++ swift/llm/sampling/vanilla_sampler.py
@@ -5,7 +5,7 @@ from copy import deepcopy
 import json
 import numpy as np
 
-from swift.experimental.sampling.base import Sampler
+from swift.llm.sampling.base import Sampler
 from swift.llm import RequestConfig
 from swift.llm.template.template_inputs import InferRequest
 from swift.utils import get_logger
